# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "TraitDataPointSchema": {
        "type": "object",
        "properties": {
          "traitName": {
            "type": "string",
            "description": "Name of the trait"
          },
          "measuredValue": {
            "type": "number",
            "nullable": true,
            "description": "Measured value from wet chemistry"
          },
          "predictedValue": {
            "type": "number",
            "nullable": true,
            "description": "Predicted value (if available)"
          },
          "year": {
            "type": "number",
            "description": "Year of measurement"
          }
        },
        "required": [
          "traitName",
          "year"
        ]
      },
      "TraitSampleSchema": {
        "type": "object",
        "properties": {
          "sampleId": {
            "type": "number",
            "description": "Unique sample ID"
          },
          "crop": {
            "type": "string",
            "nullable": true,
            "description": "Crop name"
          },
          "qualityLab": {
            "type": "string",
            "nullable": true,
            "description": "Quality lab name"
          },
          "nirModel": {
            "type": "string",
            "nullable": true,
            "description": "NIR model used"
          },
          "species": {
            "type": "string",
            "nullable": true,
            "description": "Species name"
          },
          "plantingDate": {
            "type": "string",
            "nullable": true,
            "description": "Planting date (YYYY-MM-DD)"
          },
          "physiologicalStage": {
            "type": "string",
            "nullable": true,
            "description": "Physiological stage"
          },
          "productType": {
            "type": "string",
            "nullable": true,
            "description": "Product type"
          },
          "traits": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TraitDataPointSchema"
            },
            "description": "List of traits for this sample"
          }
        },
        "required": [
          "sampleId",
          "traits"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/product-types": {
      "get": {
        "operationId": "get-product-types",
        "summary": "Get Product Types\r",
        "description": "Returns product types, optionally filtered by crop name.",
        "tags": [
          "Product-types"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Filter product types by crop name"
            },
            "required": false,
            "description": "Filter product types by crop name"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/quality-labs": {
      "get": {
        "operationId": "get-quality-labs",
        "summary": "Get Quality Labs\r",
        "description": "Returns distinct quality labs, filtered by country and/or center.",
        "tags": [
          "Quality-labs"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "country",
            "schema": {
              "type": "string",
              "description": "Country where the lab is located"
            },
            "required": false,
            "description": "Country where the lab is located"
          },
          {
            "in": "query",
            "name": "center",
            "schema": {
              "type": "string",
              "description": "Center to which the lab belongs"
            },
            "required": false,
            "description": "Center to which the lab belongs"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string",
                        "description": "Name of the quality lab"
                      },
                      "country": {
                        "type": "string",
                        "description": "Country where the lab is located"
                      },
                      "center": {
                        "type": "string",
                        "description": "Center to which the lab belongs"
                      }
                    },
                    "required": [
                      "name",
                      "country",
                      "center"
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/species": {
      "get": {
        "operationId": "get-species",
        "summary": "Get Species\r",
        "description": "Returns species, optionally filtered by crop name.",
        "tags": [
          "Species"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Filter species by crop name"
            },
            "required": false,
            "description": "Filter species by crop name"
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/traits": {
      "get": {
        "operationId": "get-traits",
        "summary": "Get Trait data per sample\r",
        "description": "Returns wet chemistry and predicted trait values,",
        "tags": [
          "Traits"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "crop",
            "schema": {
              "type": "string",
              "description": "Crop name"
            },
            "required": false,
            "description": "Crop name"
          },
          {
            "in": "query",
            "name": "qualityLab",
            "schema": {
              "type": "string",
              "description": "Quality lab name"
            },
            "required": false,
            "description": "Quality lab name"
          },
          {
            "in": "query",
            "name": "nirModel",
            "schema": {
              "type": "string",
              "description": "NIR model name"
            },
            "required": false,
            "description": "NIR model name"
          },
          {
            "in": "query",
            "name": "species",
            "schema": {
              "type": "string",
              "description": "Species name"
            },
            "required": false,
            "description": "Species name"
          },
          {
            "in": "query",
            "name": "plantingYear",
            "schema": {
              "type": "string",
              "description": "Planting year (YYYY)"
            },
            "required": false,
            "description": "Planting year (YYYY)"
          },
          {
            "in": "query",
            "name": "startDate",
            "schema": {
              "type": "string",
              "description": "Sample Date: Start date (YYYY-MM-DD)"
            },
            "required": false,
            "description": "Sample Date: Start date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "endDate",
            "schema": {
              "type": "string",
              "description": "Sample Date: End date (YYYY-MM-DD)"
            },
            "required": false,
            "description": "Sample Date: End date (YYYY-MM-DD)"
          },
          {
            "in": "query",
            "name": "physiologicalStage",
            "schema": {
              "type": "string",
              "description": "Physiological stage name"
            },
            "required": false,
            "description": "Physiological stage name"
          },
          {
            "in": "query",
            "name": "productType",
            "schema": {
              "type": "string",
              "description": "Product type name"
            },
            "required": false,
            "description": "Product type name"
          },
          {
            "in": "query",
            "name": "traitName",
            "schema": {
              "type": "string",
              "description": "Trait name(s), required. Supports comma-separated values"
            },
            "required": false,
            "description": "Trait name(s), required. Supports comma-separated values"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TraitSampleSchema"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  }
}